Clippy improvements and fixes (#26)
Turns out that the CI was not enforcing clippy lints on all features.
This caused some lints to be missed. Additionally, the stricter clippy lints (which have been part of the
`litebox` crate since approximately the beginning of this project) were
only being applied to the `litebox` crate, and not to the entire
workspace.

This PR enables the enforcement, and fixes the cases that were now being triggered.