Basic Linux Userland Runner (#82)
This PR sets up the basic userland runner. There are some more details
to be figured out, the biggest one is how we are supporting the various
syscall interception backends, but that would be separate future PRs. As
an example, currently, `litebox_platform_linux_userland` hard-commits to
the SIGSYS-based syscall interception, but this means that our
syscall-rewriter backend (which is what would be used on other
platforms) is not actually being tested.

The interface of the runner:
```console
$ cargo run --bin litebox_runner_linux_userland -- --help
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.06s
     Running `target/debug/litebox_runner_linux_userland --help`
Run Linux programs with LiteBox on unmodified Linux

Usage: litebox_runner_linux_userland [OPTIONS] <PROGRAM_AND_ARGUMENTS>...

Arguments:
  <PROGRAM_AND_ARGUMENTS>...
          The program and arguments passed to it (e.g., `python3 --version`)

Options:
      --env <ENVIRONMENT_VARIABLES>
          Environment variables passed to the program (`K=V` pairs; can be invoked multiple times)

      --forward-env
          Forward the existing environment variables

  -Z, --unstable
          Allow using unstable options

  -h, --help
          Print help (see a summary with '-h')

Unstable Options:
      --insert-file <INSERT_FILES>
          Pre-fill files into the initial file system state

      --initial-files <PATH_TO_TAR>
          Pre-fill the files in this tar file into the initial file system state

      --rewrite-syscalls
          Apply syscall-rewriter to the ELF file before running it

          This is meant as a convenience feature; real deployments would likely prefer ahead-of-time rewrite things to amortize costs.
```