Litebox allocator (#111)
This PR moves the allocator from `litebox_platform_linux_kernel` to
`litebox` so  `litebox_platform_linux_userland` can use it.

The primary reason is  to address the issue mentioned in #66, as the allocator from
libc also relies on `fs` registers to maintain per-thread metadata for
better performance. We may not need to swtich fs register anymore.