Support TLS (#156)
This PR tries to address #66. Similar to how libc uses fs on x86_64 (or
gs on x86) for thread-local storage, we could use the other one (i.e.,
gs on x86_64 and fs on x86) for litebox's TLS.

Note that on x86_64, we can directly modify fs/gs base, while on x86, we
have to call syscall `set_thread_area` to add an entry to GDT and update
`fs` selector to point to the new entry.

---------

Co-authored-by: Jay Bosamiya (Microsoft) <jayb@microsoft.com>