Fix time provider (#248)
This is part of the PR #239 that fixes time provider so that we could
convert a `Timespec` to `Instant`, which is needed for supporting futex
timeout arguments. The `std::time::Instant` does not expose this
interface and internally it uses either `Duration` (for windows) or
`Timespec` to represent it. Copied and adjusted that code from `libc`
to corresponding platform.