Add some more `Debug` impls/derives (#312)
This PR introduces `Debug` to `SyscallRequest` and related enums.

To be able to do this nicely, we actually need make a minor update the
platform interface for the raw pointer types. Specifically, we need
those pointers to support `Debug`. Not including this was partly an
oversight. Having it around makes it nicer to to debug things.