Fix tar-ro `file_status` (#316)
The `tar_ro` file-system `file_status` function would cause trouble if
it is ever passed `"/"` or `""`. @Icegrave0391 noticed that this happens
if you try to make a new directory in a place that an existing dir
already exists at a lower layer (the layered file system attempts to
check ancestors via `file_status`). This PR fixes this edge case and
adds a regression test for it.