Add a ratchet for `MaybeUninit` (#406)
`MaybeUninit` has a bunch of footguns that make it quite dangerous in
general (_technically_ it is the `assume_init` and other methods on
`MaybeUninit` that are dangerous, but I am using it as a shorthand for
its methods) and require careful review around any time it is used.

Often, zero/default-initializing a value is _very_ cheap anyways that it
is rarely worth the danger `MaybeUninit` poses.

Our codebase already has added a few `MaybeUninit`s around that I don't
think carry their own weight. Thus, this PR introduces a ratchet for
`MaybeUninit` so that we do not increase its usage, and eventually cut
down on its usage in future PRs.