LVBS: Revise per-CPU kernel variables (#369)
This PR revises the per-CPU kernel variables of the LVBS host/platform.
Now we store per-CPU kernel variables in the kernel heap (except for one
for the BSP), allowing us to support a large number of VCPUs. These
per-CPU kernel variables can be accessed through the GS registers using
`with_per_cpu_variables` and a closure along with `RefCell` to avoid
multiple mutable borrows. In addition, this PR lets us build LiteBox for
LVBS with the release profile.