Clear out manually-handled `File` descriptors from shim (#408)
This PR clears out the manually handled `File` Descriptors from the
shim. It also updates the stdio initialization to do it eagerly rather
than on-demand. The moment we have an FS, we can initialize FDs 0, 1,
and 2.

---

This is yet another PR in the series of PRs towards the better FD design
 (connected to #31).