Set up pipes as an FD-enabled subsystem (#446)
This PR makes `litebox::pipes` an FD-enabled subsystem.

This specific PR is concerned entirely with introducing a new public
interface to be used; a future PR will actually use this from the shim.

This shim updates the tests _within_ the LiteBox crate itself to use
this new interface, so that a future PR can eliminate the direct
dependence on internal details more cleanly.

---

This is yet another PR in the series of PRs towards the better FD design
(connected to https://github.com/microsoft/litebox/issues/31).