No vdso on rewriter backend (#469)
Now that we have enough support for syscalls, the Linux runner now can
run without `vdso`, thus we don't need the patch on `aux` vec to pass
CI.

Another issue with `vdso` I found recently was that `vdso` on x86 has
this function `kernel_vsyscall` that allow guest to call any syscalls,
but our rewriter backend does not instrument the `vdso` section. To
address it, we don't provide `vdso` address to guest if it runs with the
rewriter backend. This ensures that all syscalls would go through
LiteBox.