More thread startup cleanup (#509)
Make Linux shim entrypoint objects to be per-thread and not global, and
update the platforms to handle this. This eliminates the global shim
registration, which in theory would allow multiple shims to be used at
the same time (on different threads, of course).

For now, there is still a bit of a TLS dance, which should go away once
we stop using TLS in the Linux shim.

Also, stop returning contexts from the shim during program load and
instead initialize the context during the first call into the shim. This
will make switching to platform-specific contexts easier in the future.

The OPTEE shim keeps its current model for now.