Unify and improve userspace pointers (#516)
* Use a single userspace pointer type for all the platforms.
* Add a trait to allow the platforms to perform pointer validation (i.e.
to ensure that the pointer is not in the kernel half of the address
space), but do not implement any validation yet.
* Extend the exception table functionality to support Windows userspace,
borrowing from the `trycopy` crate from OpenVMM.
* Use inline asm instead of a separate file to improve code generation.
* Use ordinary `mov`s for small accesses, guaranteeing that small,
aligned accesses are atomic. Use this to fix the undefined behavior in
the futex manager.

Future directions:
* Implement the appropriate validation for the kernel platforms.
* Add mechanisms to support SMAP.
* Improve memcpy performance--`rep movsb` is very slow for small
accesses on many processors, even some new ones.